requirements.txt
pyyaml
pytest
anthropic
python-dotenv
